+++ student-distrib/process.h
@@ -15,7 +15,7 @@
 // pcb_t* get_pcb(); //return the current operating pcb based on curr_process global watcher
 // pcb_t* get_pcb_from_pid(uint32_t pid); //return the pcb based on the given pid
 int32_t system_execute(const uint8_t* command);
-int32_t system_halt(uint8_t status);
+int32_t system_halt(uint16_t status);
 pcb_t* getRunningPCB(); //get the most recently running pcb from the stack
 // void iret_helper();
